+++ lldb/source/API/SBTarget.cpp
@@ -13,6 +13,7 @@
 #include "lldb/API/SBEvent.h"
 #include "lldb/API/SBExpressionOptions.h"
 #include "lldb/API/SBFileSpec.h"
+#include "lldb/API/SBLineEntry.h"
 #include "lldb/API/SBListener.h"
 #include "lldb/API/SBModule.h"
 #include "lldb/API/SBModuleSpec.h"
@@ -1685,6 +1686,27 @@ SBSymbolContextList SBTarget::FindCompileUnits(const SBFileSpec &sb_file_spec) {
   if (TargetSP target_sp = GetSP(); target_sp && sb_file_spec.IsValid())
     target_sp->GetImages().FindCompileUnits(*sb_file_spec, *sb_sc_list);
   return sb_sc_list;
+}
+
+SBSymbolContextList
+SBTarget::FindSymbolContexts(const SBLineEntry &line_entry,
+                             lldb::SymbolContextItem resolve_scope,
+                             bool check_inlines) {
+  LLDB_INSTRUMENT_VA(this, line_entry, resolve_scope, check_inlines);
+
+  SBSymbolContextList sc_list;
+  TargetSP target_sp = GetSP();
+  if (!target_sp || !line_entry.IsValid())
+    return sc_list;
+
+  SBFileSpec file_spec = line_entry.GetFileSpec();
+  if (!file_spec.IsValid())
+    return sc_list;
+
+  // TODO: Pass column when ResolveSymbolContextsForFileSpec supports it.
+  target_sp->GetImages().ResolveSymbolContextsForFileSpec(
+      *file_spec, line_entry.GetLine(), check_inlines, resolve_scope, *sc_list);
+  return sc_list;
 }
 
 lldb::ByteOrder SBTarget::GetByteOrder() {